Set iocage props on root dataset
